fix(tests): 🐛 ensure unique free port for HttpListener
