+++ DomainDetective.Tests/TestPlainHttpHealthCheck.cs
@@ -53,11 +53,7 @@ namespace DomainDetective.Tests {
         }
 
         private static int GetFreePort() {
-            var listener = new System.Net.Sockets.TcpListener(IPAddress.Loopback, 0);
-            listener.Start();
-            var port = ((IPEndPoint)listener.LocalEndpoint).Port;
-            listener.Stop();
-            return port;
+            return PortHelper.GetFreePort();
         }
     }
 }
